+++ tests.py
@@ -25,6 +25,8 @@ from unittest.mock import patch
 import compatibility
 import pytest
 
+from compatibility import err
+
 
 def test_missing_or_empty_paramameters():
     "3 parameters are required, the other 3 have defaults."
@@ -103,13 +105,13 @@ def test_release_date():
         package_version='0.1',
         release_date='2021-01-01')
     # malformed date string
-    with pytest.raises(ValueError):
+    with pytest.raises(err.BadDate):
         compatibility.Check(
             package_name='test',
             package_version='0.1',
             release_date='2021-Jan-10')
     # valid string format, but invalid date
-    with pytest.raises(ValueError):
+    with pytest.raises(err.BadDate):
         compatibility.Check(
             package_name='test',
             package_version='0.1',
@@ -370,7 +372,7 @@ def test_check_system_CONTRADICTIONS():
     with patch('platform.system') as system:
         system.return_value = 'Windows'
         # Cannot be incompatible and have full support
-        with pytest.raises(ValueError) as excinfo:
+        with pytest.raises(err.ParameterContradition) as excinfo:
             compatibility.Check(
                 package_name='test',
                 package_version='1',
@@ -380,7 +382,7 @@ def test_check_system_CONTRADICTIONS():
                 )
         assert 'support AND be incompatible' in str(excinfo.value)
         # cannot be fully and partialy supported
-        with pytest.raises(ValueError) as excinfo:
+        with pytest.raises(err.ParameterContradition) as excinfo:
             compatibility.Check(
                 package_name='test',
                 package_version='1',